starting user encryption flow
