+++ api.py
@@ -4,8 +4,7 @@ from tornado import options
 
 from app.user_auth import UserAuth
 from app.user_deauth import UserDeauth
-from app.ticket import TicketHandler
-from app.showtimes import CreateShowtimeHandler
+from app.showtimes import CreateShowtimeHandler, ListShowtimesHandler
 from app.services_auth.google_auth import GoogleAuth
 from app.services_auth.facebook_auth import FacebookAuth
 from app.services_auth.spotify_auth import SpotifyAuth
@@ -93,7 +92,7 @@ if __name__ == "__main__":
             (r'/auth/instagram', InstagramAuth),
             (r'/auth/close', CloseWindow),
 
-            (r'/tickets', TicketHandler),
+            (r'/api/showtimes', ListShowtimesHandler),
             (r'/api/showtimes/create', CreateShowtimeHandler),
             (r"/favicon.ico", web.StaticFileHandler, {"path": ""}),
         ],
